Set GROOVY_HOME so that startGroovy does not fail.
/usr/lib/bigtop-groovy/bin/startGroovy fails when trying to detect
GROOVY_HOME.

We start the init-hdfs.sh script in /home/vagrant. The script sudos
to hdfs, then calls startGroovy. startGroovy then tries to detect
GROOVY_HOME in /usr/lib/... with `cd`, then `cd` back to where it
was before. That fails because /home/vagrant (where it was before)
is chmod'd 700.